Move CePO into its own directory
